refactor: Rename the components in config form
