+++ packages/platform/src/containers/NotificationConfiguration/ConfigForm/components/BoxInput.tsx
