+++ packages/platform/src/containers/NotificationConfiguration/ConfigForm/components/List/ListItem.tsx
